team10/requirements.txt
whitenoise
jdatetime
requests
djangorestframework>=3.14.0
python-dateutil>=2.8.2